post/serializers.py
from rest_framework import serializers
from .models import Post,Feedback

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ('title', 'tag', 'description','time','slug')

class FeedbackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Feedback
        fields = '__all__'