requirements.txt
django == 3.0.4
djangorestframework == 3.11.0
django-ckeditor == 5.8.0
django-cors-headers == 3.2.1